esp32: a platform, and a repl on real hardware
ESP-IDF plays the role UEFI plays elsewhere, so this platform is shaped
like efi's rather than microvm's: the firmware surface kernel.c uses is
small, and IDF answers all of it. Boots to an interactive repl on a
LilyGo T-Deck and on an M5Stack Cardputer, and runs the boot tests
under Espressif's qemu.

Two build systems, because IDF's link is generated linker fragments,
section placement, keep-lists, init arrays, an app descriptor and a
bootloader image -- not a list of libraries meson could be handed.
tools/sources.lua is the seam that keeps them honest: meson and IDF's
cmake both read it, so a new source cannot reach one build and miss the
other. esp32/meson.build is a sibling project rather than a branch in
the top-level one; threading an "if platform == esp32" through a
thousand lines that all assume they produce an image would risk every
working platform to gain nothing.

kernel.h declared kernel_spawn_buffer taking unsigned long where
kernel.c defines size_t. Identical under LP64 and a hard error on any
32-bit target -- the only error in 8153 lines of arch-blind C compiled
for xtensa.

The board is a build-time choice (Kconfig.projbuild, default NONE)
because a probe for absent hardware does not merely answer no: the
T-Deck SD probe powers a rail and initialises SPI2, and under qemu it
never returned. The machine hung before cons and every boot test timed
out, unnoticed, because a hang and a slow boot look identical.

Devices: the microSD as raw sectors, deliberately not
esp_vfs_fat_sdspi_mount -- that mounts FAT behind newlib's fopen,
outside the per-proc namespace, where no proc can be granted or denied
it. sdmmc_read_sectors is the surface EFI_BLOCK_IO and virtio_blk
already give, so lib/gpt.lua and lib/gefs go above it unchanged. And
the Cardputer keyboard, which cannot interrupt: a 74HC138 drives one
column low and seven actively high, so no key outside the selected
column can pull a row down. It is scanned from the idle path instead,
rate limited to 8ms -- a scan is 56.6us measured, and the shim's wait
loop turns over every 1ms.

Four traps, all silent:

  - luaL_loadfile uses fopen, which here is newlib's and goes to IDF's
    VFS rather than src/libc. Everything failed with "No such file or
    directory" -- an errno string, which is the tell that the request
    never reached our tree. vfs.c registers the embedded set instead.
  - embed_load hands out a malloc'd copy and lua opens every module
    twice, so a 44KB module cost 88KB per require and the third open
    failed for want of memory, reported as a missing file.
  - console_write is ESP-IDF's symbol too, and cmake splits
    LUA_PATH_DEFAULT on its semicolon into a second -D. Both are solved
    by a force-included prefix.h, so platform.h and the other three
    platforms are untouched.
  - USB-Serial-JTAG writes without the driver but cannot read, so the
    console looks perfect and only typing is dead.

The console is where the board and the emulator genuinely disagree, and
no single setting serves both: sdkconfig.qemu overrides the board's for
qemu, where a USB-primary image boot-loops with no output at all.

+++ src/kernel.h
@@ -5,6 +5,8 @@
  * queues, rights are per-proc handles onto ports.
  */
 
+#include <stddef.h>
+
 /* the TSC epoch and rate. must run before anything measures a duration
  * or logs a line; efi_main calls it first. costs 100ms of BS->Stall.
  */
@@ -18,7 +20,7 @@ void	kernel_log(const char *s);
 
 int	kernel_init(void);
 int	kernel_spawn_file(const char *path);	/* returns pid or -1 */
-int	kernel_spawn_buffer(const char *code, unsigned long len);
+int	kernel_spawn_buffer(const char *code, size_t len);
 void	kernel_run(void);			/* until all procs die */
 
 /* checked capability, write/append only (read is ambient): does
